feat: added MIT License
Introduced the MIT License to the project to clarify usage rights, modification, distribution, and liability terms for all users. This license supports open collaboration by allowing others to freely share, modify, and enhance the software while ensuring that the original copyright remains noted. This step formalizes the framework under which contributions are made, encouraging a broader use and contribution from the community, ensuring the software remains free and open for everyone’s benefit.